Replace 2dtree w/ 3d for xmatch
Old implementation is still available in kdtree_gnomonic_match
